address copilot review comments
- include cstdlib in the unit test to ensure that free prototype is
  available
- change a loop variable from int to size_t to avoid warnings comparing
  int's to size_t's
- mark description and is_virtual_p2p_link as mutable in the PCIeNode
  structure to avoid undefined behavior when changing those values after
  inserting the structure into a std::set

+++ projects/rocshmem/src/gda/topology.cpp
@@ -739,7 +739,7 @@ namespace rocshmem
     std::set<int> matches = {};
 
     // Loop over the candidates to find the ones with the lowest common ancestor (LCA)
-    for (int i = 0; i < candidateBusIdList.size(); i++) {
+    for (size_t i = 0; i < candidateBusIdList.size(); i++) {
       std::string const& candidateBusId = candidateBusIdList[i];
       if (candidateBusId == "") continue;
       PCIeNode const* lca = GetLcaBetweenNodes(root, targetBusId, candidateBusId);
